pyproject.toml
[tool.poetry]
name = "musicbucket"
version = "0.1.0"
description = ""
authors = ["Pau Riera <pau.riera.forteza@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "==4.0"
djangorestframework = "==3.*.*"
django-compressor = {git = "https://github.com/django-compressor/django-compressor.git", rev = "develop"}  # TODO: Change to 3.0 after 2022/01/01
django-constance = {version = "==2.8.0", extras = ["database"]}
django-storages = "==1.12.3"
django-extensions = "==3.1.5"
django-kaio = "==0.14.3"
django-robots = {git = "https://github.com/jazzband/django-robots.git"}
django-widget-tweaks = "==1.4.9"
pytz = "==2021.3"
boto3 = "==1.20.*"
django-bootstrap4 = "==21.1"
django-crispy-forms = "==1.13.0"
psycopg2-binary = ">=2.8,<2.9"
django-redis = "==5.1.0"
emoji = "==1.6.1"
markdown = "==3.3.6"
django-filter = "==21.1"
spotipy = "==2.19.0"
pylast = "==4.4.*"
musicbrainzngs = "==0.7.1"
sentry-sdk = "==1.5.*"
raven = "==6.10.0"
django-telegram-login = "==0.2.3"
django-tables2 = "==2.4.1"
gunicorn = "==20.1.0"
django-hijack = "==3.1.4"
django-cookie-law = "==2.1.0"
whitenoise = "==5.3.0"
lastfmcollagegenerator = "==0.1.2"
django-cors-headers = "==3.*.*"
django-rq = "==2.5.1"
djangorestframework-simplejwt = "==5.*.*"


[tool.poetry.dev-dependencies]
pytest = "*"
pytest-checkipdb = "*"
pytest-django = "*"
pytest-eradicate = "*"
pytest-flake8 = "*"
pytest-splinter = "*"
pytest-sugar = "*"
pytest-xdist = "*"
django-debug-toolbar = "*"
ipdb = "*"
whitenoise = "*"
flake8-django = "*"
model-bakery = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
